scss changes and readme update

+++ src/js/date.js
@@ -1,6 +1,6 @@
 // Skapar ett Date objekt - current day and date
 var date = new Date(),
-options = { weekday: "long", month: "short", day: "numeric", year: "numeric" };
+options = { weekday: "long", month: "long", day: "numeric", year: "numeric" };
 
 // Konverterar och lägger till språk/land
 date = date.toLocaleString( "sv-SE", options );
